+++ azalea/src/auto_tool.rs
@@ -4,7 +4,7 @@ use azalea_entity::{ActiveEffects, Attributes, FluidOnEyes, Physics, inventory::
 use azalea_inventory::{ItemStack, Menu, components};
 use azalea_registry::builtin::{BlockKind, EntityKind};
 
-use crate::Client;
+use crate::{Client, client_impl::error::AzaleaResult};
 
 #[derive(Debug)]
 pub struct BestToolResult {
@@ -13,7 +13,7 @@ pub struct BestToolResult {
 }
 
 impl Client {
-    pub fn best_tool_in_hotbar_for_block(&self, block: BlockState) -> BestToolResult {
+    pub fn best_tool_in_hotbar_for_block(&self, block: BlockState) -> AzaleaResult<BestToolResult> {
         self.query_self::<(
             &Inventory,
             &Physics,
@@ -35,15 +35,16 @@ impl Client {
         )
     }
 
-    pub async fn mine_with_auto_tool(&self, block_pos: BlockPos) {
+    pub async fn mine_with_auto_tool(&self, block_pos: BlockPos) -> AzaleaResult<()> {
         let block_state = self
-            .world()
+            .world()?
             .read()
             .get_block_state(block_pos)
             .unwrap_or_default();
-        let best_tool_result = self.best_tool_in_hotbar_for_block(block_state);
+        let best_tool_result = self.best_tool_in_hotbar_for_block(block_state)?;
         self.set_selected_hotbar_slot(best_tool_result.index as u8);
         self.mine(block_pos).await;
+        Ok(())
     }
 }