+++ azalea/src/lib.rs
@@ -47,7 +47,15 @@ pub use azalea_registry as registry;
 #[doc(hidden)]
 #[deprecated(note = "renamed to `Identifier`.")]
 pub type ResourceLocation = azalea_registry::identifier::Identifier;
-pub use azalea_chat as chat;
+
+// TODO: replace this mod with the commented line below
+// pub use azalea_chat as chat;
+pub mod chat {
+    pub use azalea_chat::*;
+    #[deprecated = "moved to `azalea::client_chat`."]
+    pub type ChatPacket = azalea_client::client_chat::ChatPacket;
+}
+
 pub use azalea_registry::identifier::Identifier;
 pub use azalea_world as world;
 pub use bevy_app as app;