+++ CHANGELOG.md
@@ -13,6 +13,7 @@ is breaking anyways, semantic versioning is not followed.
 ### Changed
 
 - Re-export `azalea-chat` from `azalea`, and move the other `azalea::chat` module to `azalea::client_chat`.
+- Many functions in `Client` and `EntityRef` now return an `AzaleaResult` instead of panicking when getting entity data fails.
 
 ### Fixed